src/test/java/it/madlabs/patternrec/web/rest/service/PointServiceTest.java
package it.madlabs.patternrec.web.rest.service;

import it.madlabs.patternrec.web.rest.model.Point;
import it.madlabs.patternrec.web.rest.model.Line;
import it.madlabs.patternrec.web.rest.repository.PointRepository;
import org.junit.*;

import java.math.BigDecimal;
import java.util.Arrays;
import java.util.LinkedList;
import java.util.List;
import java.util.Random;

public class PointServiceTest {

    //Better to use some mock frameworks and implements interface
    private class MockRepository extends PointRepository{
        @Override
        public List<Point> findAll() {
            return Arrays.asList(new Point(2d,3d), new Point(-2d,1023d), new Point(3.2d,0d), new Point(1d,1d));
        }
    }

    private PointService pointService = new PointService(new MockRepository());

    @Before
    public void setUp() throws Exception {
    }

    @After
    public void tearDown() throws Exception {
    }

    @Test(expected = IllegalArgumentException.class)
    public void given_null_point_set_when_call_addPoint_then_should_throw_IllegalArgumentsException(){
        pointService.addPoint(null);
        Assert.fail();
    }

    @Test
    public void given_empty_point_set_when_call_addPoint_then_should_return_point_set_with_added_element(){
        Point point = new Point(1d, 1d);
        pointService.addPoint(point);
        List<Point> pointSet = pointService.allPoints();
        Assert.assertEquals(1, pointSet.size());
        Assert.assertEquals(point, pointSet.get(0));
    }

    @Test
    public void given_NOT_EMPTY_point_set_when_call_addPoint_with_just_added_point_then_should_NOT_recalculate_lines(){
        pointService.loadPoints();
        List<List<Point>> collinearPoints = pointService.allCollinearPoints(2);
        int before = collinearPoints.size();
        Point point = new Point(1d, 1d);
        pointService.addPoint(point);
        collinearPoints = pointService.allCollinearPoints(2);
        int after = collinearPoints.size();
        Assert.assertEquals(before, after);
    }

    @Test
    public void given_NOT_EMPTY_point_set_when_call_addPoint_with_new_point_then_should_recalculate_lines(){
        pointService.loadPoints();
        Point point = new Point(3d, 3d);
        pointService.addPoint(point);
        List<List<Point>> collinearPoints = pointService.allCollinearPoints(2);
        Assert.assertNotEquals(0, collinearPoints.size());
    }

    @Test
    public void given_2_collinear_point_set_when_call_allCollinearPoint_then_should_return_2_points(){
        pointService.deleteAllPoints();
        pointService.addPoint(new Point(1d, 1d));
        pointService.addPoint(new Point(2d, 2d));
        List<List<Point>> collinearPoints = pointService.allCollinearPoints(2);
        Assert.assertEquals(1, collinearPoints.size());
        Assert.assertEquals(2, collinearPoints.get(0).size());
    }


    @Test
    public void given_3_collinear_point_set_when_call_allCollinearPoint_then_should_return_3_points(){
        pointService.deleteAllPoints();
        pointService.addPoint(new Point(1d, 1d));
        pointService.addPoint(new Point(2d, 2d));
        pointService.addPoint(new Point(3d, 3d));
        List<List<Point>> collinearPoints = pointService.allCollinearPoints(3);
        Assert.assertEquals(1, collinearPoints.size());
        Assert.assertEquals(3, collinearPoints.get(0).size());
    }

    @Test
    public void given_4_collinear_point_set_when_call_allCollinearPoint_then_should_return_4_points(){
        pointService.deleteAllPoints();
        pointService.addPoint(new Point(1d, 1.825d));
        pointService.addPoint(new Point(2d, 3.025d));
        pointService.addPoint(new Point(100d, 120.625d));
        pointService.addPoint(new Point(-100d, -119.375d));
        List<List<Point>> collinearPoints = pointService.allCollinearPoints(4);
        Assert.assertEquals(1, collinearPoints.size());
        Assert.assertEquals(4, collinearPoints.get(0).size());
    }

    @Test
    public void given_X_calculated_collinear_point_set_when_call_allCollinearPoint_then_should_return_X_points(){
        final int NUMBER_OF_POINT = 100;
        Random random = new Random();
        BigDecimal a = new BigDecimal(random.nextInt(10000));
        BigDecimal c = new BigDecimal(random.nextInt(10000));
        Line randomLine = new Line(a, BigDecimal.ONE, c);
        pointService.deleteAllPoints();
        for (int i = 0; i < NUMBER_OF_POINT; i++) {
            Double x = new Double(i);
            Double y = randomLine.calculateYgivenX(x);
            pointService.addPoint(new Point(x, y));
        }
        List<List<Point>> collinearPoints = pointService.allCollinearPoints(NUMBER_OF_POINT);
        Assert.assertEquals(1, collinearPoints.size());
        Assert.assertEquals(NUMBER_OF_POINT, collinearPoints.get(0).size());
    }

//  It wont work: I think that the problem is conversion from double to BigDecimal and vice versa
//    @Test
//    public void given_X_calculated_collinear_point_set_when_call_allCollinearPoint_then_should_return_X_points(){
//        final int NUMBER_OF_POINT = 100;
//        Random random = new Random();
//        BigDecimal a = new BigDecimal(random.nextInt(10000));
//        BigDecimal c = new BigDecimal(random.nextInt(10000));
//        Line randomLine = new Line(a, BigDecimal.ONE, c);
//        pointService.deleteAllPoints();
//        for (int i = 0; i < NUMBER_OF_POINT; i++) {
//            Double x = new Double(random.nextDouble());
//            Double y = randomLine.calculateYgivenX(x);
//            pointService.addPoint(new Point(x, y));
//        }
//        List<List<Point>> collinearPoints = pointService.allCollinearPoints(NUMBER_OF_POINT);
//        Assert.assertEquals(1, collinearPoints.size());
//        Assert.assertEquals(NUMBER_OF_POINT, collinearPoints.get(0).size());
//    }

}